requirements.txt
scikit-learn
streamlit
matplotlib
pandas
scikit-image
opencv-python
pillow
numpy
pickle
